Refactor Welcome component & tests to simplify code
